pyproject.toml
[project]
name = "biothrone"
version = "0.0.1"
description = "An app that helps you achieve financial freedom"
authors = [
    {name = "Jun Siang Pang",email = "junsiang97@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "finnhub-python (>=2.4.26,<3.0.0)",
    "alpha-vantage (>=3.0.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fredapi (>=0.5.2,<0.6.0)",
    "requests (>=2.32.5,<3.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "massive (>=2.0.2,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "edgartools (>=5.7.1,<6.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langchain (>=1.2.6,<2.0.0)",
    "langchain-deepseek (>=1.0.1,<2.0.0)",
    "deepagents (>=0.4.4,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
